### 2024-10-02 #7223 @ivastar
@enthusi I will be taking over as editor of this submission. I will be looking for reviewers for the next week or two. I see you have already made recommendations for reviewers above, thank you! In the mean time, the Statement of Need is one of the [required sections](https://joss.readthedocs.io/en/latest/review_checklist.html) in the manuscript:

A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Could you please add it?
