### 2024-10-02 f6a7723
adds Statement of need to the JOSS paper
