### 2024-11-01 #7223 @enthusi
We have an accepted yet unpublished survey paper which used Feline extensively and it would be great if we could link to the JOSS entry once available. Is there any planned schedule for the review (clearly just a lower limit, not upper limit)?
