### 2024-11-05 #7223 @enthusi
Oh, if it helps for finding reviewers: the actual code part is in C/OpenMP and not very complex.
The Python code is for minor preprocessing and otherwise merely for plotting the outcome.
The CUDA is fully optional and just replaces the main C part, so it is not even required at all.
Thanks alot for your efforts and reply.
