+++ joss/paper.bib
@@ -142,3 +142,19 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{2024arXiv241107014B,
+       author = {{Bouche2024}, N.~F. and {Wendt}, M. and {Zabl}, J. and {Cherrey}, M. and {Schroetter}, I. and {Langan}, I. and {Muzahid}, S. and {Schaye}, J. and {Epinat}, B. and {Wisotzki}, L. and {Contini}, T. and {Richard}, J. and {Bacon}, R. and {Weilbacher}, P.~M.},
+        title = "{MusE GAs FLOw and Wind (MEGAFLOW) XII. Rationale and design of a MgII survey of the cool circum-galactic medium with MUSE and UVES: The MEGAFLOW Survey}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Astrophysics of Galaxies},
+         year = 2024,
+        month = nov,
+          eid = {arXiv:2411.07014},
+        pages = {arXiv:2411.07014},
+archivePrefix = {arXiv},
+       eprint = {2411.07014},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024arXiv241107014B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
