### 2024-11-13 #7223 @enthusi
I just updated the JOSS paper (merely adding a reference to the mentioned survey paper) in the repository.
