+++ joss/paper.bib
@@ -142,8 +142,8 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{2024arXiv241107014B,
-       author = {{Bouche2024}, N.~F. and {Wendt}, M. and {Zabl}, J. and {Cherrey}, M. and {Schroetter}, I. and {Langan}, I. and {Muzahid}, S. and {Schaye}, J. and {Epinat}, B. and {Wisotzki}, L. and {Contini}, T. and {Richard}, J. and {Bacon}, R. and {Weilbacher}, P.~M.},
+@ARTICLE{Bouche2024,
+       author = {{Bouch{\'e}}, N.~F. and {Wendt}, M. and {Zabl}, J. and {Cherrey}, M. and {Schroetter}, I. and {Langan}, I. and {Muzahid}, S. and {Schaye}, J. and {Epinat}, B. and {Wisotzki}, L. and {Contini}, T. and {Richard}, J. and {Bacon}, R. and {Weilbacher}, P.~M.},
         title = "{MusE GAs FLOw and Wind (MEGAFLOW) XII. Rationale and design of a MgII survey of the cool circum-galactic medium with MUSE and UVES: The MEGAFLOW Survey}",
       journal = {arXiv e-prints},
      keywords = {Astrophysics - Astrophysics of Galaxies},
