### 2024-11-25 #7528 @editorialbot
[REVIEW]: FELINE: A tool to detect emission line galaxies in 3d data
**Submitting author:** @enthusi (<a href="http://orcid.org/0000-0001-5020-9994">Martin Wendt</a>)
**Repository:** https://github.com/enthusi/feline
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.0
