### 2024-12-15 #7528 @smutch
> @smutch, @bronreichardtchu since the issue has been quiet for a couple of weeks, sending a ping to remind you of this review. Please let me know if there is any way I can support your effort.

Apologies for the delay. I'll try to have my review by the end of the week at the latest.
