### 2024-12-17 49b8125
change expression feline.c
thanks for the hint.
resolves #8
