### 2024-12-17 enthusi/feline#9 @bronreichardtchu
Typo in Usage Guide in README
I've been following through the instructions for manual usage.  When I get to this step:
https://github.com/enthusi/feline/blob/49b81256574f5a3b47eb693418af0e7b923ab053/README.md?plain=1#L84-L87

it gives me the error:
```
[Errno 2] No such file or directory
```

I looked for the corresponding line in your makefile, and I think you need to change this code to cd into the preprocessing folder, not the postprocessing folder.
