### 2024-12-17 enthusi/feline#10 @bronreichardtchu
Issue with make: nproc on macOS
I might be wrong about this, since I am primarily a python user, and don't usually have to interact with make.  However.  I am running this on a Mac, and having issues with the following lines:
https://github.com/enthusi/feline/blob/49b81256574f5a3b47eb693418af0e7b923ab053/README.md?plain=1#L89-L94

when I run make, I get the error:
```
make: nproc: Command not found
gcc -O3 -ffast-math -fopenmp -g -std=c99  -D SDLavailable=0 src/feline.c -o feline.bin -lm
clang: error: unsupported option '-fopenmp'
clang: error: unsupported option '-fopenmp'
make: *** [feline.bin] Error 1
```

I've done some googling, and I don't think nproc is a command available on macOS. Is there an easy workaround, and can it be included in the README?
