### 2024-12-17 6824301
fix typo README.md
thank you for the hint.
resolves #9
