### 2024-12-19 #7528 @smutch
# Review

I've had a good look through the repo and successfully run the non-CUDA version on my laptop (confirmation of successfully running the CUDA version to follow).

I think it's great that you've made this code available for the community. It's really  important for reproducibility and great that you've taken the time and effort to do this. However, I do have a number of suggestions, some of which I think need to be addressed to meet the requirements of the journal.

The main things which I think need to be addressed is the lack of significant documentation and, potentially, more useful tests. I would also suggest that the removal of duplicate code and a general clean up of the repository is required.

### 1. General checks

1. In general, the codebase could do with a tidy up. The are a number of files in `src/preprocessing/lib` which appear to be reproduced in `src/preprocessing` with little modification. It looks as if some files in `src/preprocessing` import from `lib` whilst others don't, meaning there are two almost identical versions of the same file being imported. Have I understood this correctly?

2. Given that it appears that a large amount (~3.2k lines?) of code is re-vendored from LSDcat I think it would be good to indicate somewhere what the key differences made to the LSDcat code are here (if any). If there is very little modification to LSDcat, I would recommend explicitly noting in the paper that LSDcat is used to carry out the preprocessing. Also - is there a particular reason why LSDcat is vendored here rather than linked as a git submodule, for example?

3. It appears that the code is meant to be run on Linux only, but that is not noted anywhere. I had to make minor modifications to the Makefile to get it to run on my Mac.

### 2. Functionality

1. _Minor suggestion:_ There are some intermediate outputs generated in the source directory. It's great that there is a `clean` recipe in the Makefile but, if the you're willing, I would suggest that output should be clearly separated from the code so that it can be recorded and versioned by the user more easily.

### 3. Documentation

1. The only output noted by the readme is the PDF. Is this the only output that would be of interest to a potential user? What if they wanted the pixel masks and spectral line detections so that they can pass them through their own subsequent analysis? Documentation on where this information can be found and how the user can read it would be useful.
2. _Minor suggestion:_ The docstrings at the top of the `src/postprocessing` files are clearly generated by AI. I have nothing against this, but I would definitely suggest cleaning them up and making them more useful and less verbose.
3. There is no documentation on how to interpret the PDF. What are the "collapsed", "white", "quality" and "no_lines" panels? What does the colour coding of the spectral lines mean? How should "match_strength" be interpreted? What is a good/bad value for this? Also, there are no units on any of the axes. A sphinx (or similar) documentation site with an example run through of the code and interpretation of the outputs would be one way to address this and point 1 above.
4. There are automated tests and CI to verify some basic functionality of the software. There are no instructions for the user on how to run these though. From looking at the CI, it looks like a specific mock fits file is required to run the tests and is downloaded from another repository? As a minimum, I would recommend you document the full testing procedure and ensure that users have instructions to access to the required input data.
5. It's hard to tell, but it appears (and I'm sorry if I've got this wrong) that the tests are relatively basic (i.e. they test that a file is written or a function runs without error). It doesn't appear as if the actual main functionality of the package is tested. i.e. Are the correct lines detected with the correct metrics? Are the correct sources detected? etc.

### 4. Software paper

1. There is no description of how the software differs from other commonly used packages in the area. Is that because there aren't any viable alternatives?
