### 2024-12-19 #7528 @ivastar
> 3\. It appears that the code is meant to be run on Linux only, but that is not noted anywhere. I had to make minor modifications to the Makefile to get it to run on my Mac.

@smutch thank you for the detailed review! To your point above, since you were able to get this to run on a Mac, would you be able to make a pull request adding the edited Makefile for Mac users?
