### 2024-12-19 #7528 @enthusi
Thank you for effort @smuth.
It supposedly runs on all three main OS'es, but we lacked some testing abilities. The issue reported should have the work around of using 'sysctl -n hw.ncpu' instead of nproc and we will incorporate this. We are not yet aware of other problems.
Regarding the general deficiency of in-depth documentation, we were under the assumption to limit ourselves to the JOSS journal text limit itself. I am happy to add more details on the process in the repository instead. In the coming days/weeks we will work through all of your helpful suggestions. A quick reply to point #4: LSDcat is the only tool that applies the 3D template single line matching to cubic data. FELINE is the only tool that detects whole galaxies with that approach and it has been put to use in the recent paper https://arxiv.org/abs/2411.07014. We might still make it in time to appear in the references of the proof, if we managed to polish the JOSS submission accordingly.
