### 2024-12-20 enthusi/feline#11 @bronreichardtchu
Typo in README manual usage instructions
I think this line should not have a space in it after the `./`
https://github.com/enthusi/feline/blob/68243010ab2bc18a2740a4bcdcb2422fa988b95a/README.md?plain=1#L93
