### 2025-01-02 enthusi/feline#12 @bronreichardtchu
Add to instructions
Another small thing - you should make it clear in the README instructions in the lines below that you need to`cd ../..` back to the main directory before running the `make clean` command

https://github.com/enthusi/feline/blob/68243010ab2bc18a2740a4bcdcb2422fa988b95a/README.md?plain=1#L112-L115
