### 2025-01-02 enthusi/feline#13 @bronreichardtchu
Broken link 
Just noticed there's a link broken here:

https://github.com/enthusi/feline/blob/68243010ab2bc18a2740a4bcdcb2422fa988b95a/CONTRIBUTING.md?plain=1#L101

Probably worth fixing while cleaning everything up!
