### 2025-01-03 #7528 @bronreichardtchu
## Review
I've successfully run the non-CUDA version on my laptop.  I found the instructions in the README for both the step-by-step and automatic usage generally easy to follow, once I'd worked through the MacOS specific problems.

The idea for this code is really great, and one that is definitely needed.  However, in order for it to be more easily taken up by the community and meet the requirements of the journal, I think that more documentation of the code is needed, particularly around what the options for running it are/mean, and what the outputs and results are.
### General Checks
1. I agree with @smutch - I noticed that `lsd_cat_lib.py` and `line_em_funcs.py` are included twice - once in `preprocessing/` and once in `preprocessing/lib`. From a quick look, these appear to have exactly the same code (although I might have missed some tiny changes). Consider deleting one version of these files to avoid ending up with two versions of the same code.
2. I had two main problems with making the code run on my Mac -
	1. Firstly, `nproc` is not available on macOS.  I solved this by simply inputting the number of cores available in the makefile, but there should be a more general answer to this (I opened an issue here: https://github.com/enthusi/feline/issues/10).
	2. Secondly, I needed to change the CC to clang instead of gcc.  I think this could be done with an if statement in the makefile as well.
3. I tested this with the two versions of Python I have available on my laptop - Python 3.10.12 worked as expected.  Python 3.13.1 did not work, there were issues with the numpy imports, but this might be another issue specific to MacOS (there's a similar import issue reported here: https://trac.macports.org/ticket/71327).
### Functionality
1. *Minor suggestion:* I would suggest adding a copiable line under the Installation Instructions heading in the README on how to git clone the code.  This is purely to make it extra accessible to people just starting out.
### Documentation
 1. I found the step-by-step instructions on how to run the code really helpful.  I've submitted a couple of issues (https://github.com/enthusi/feline/issues/9; https://github.com/enthusi/feline/issues/11; https://github.com/enthusi/feline/issues/12; https://github.com/enthusi/feline/issues/13) where I found typos or missed instructions as I was running things.  I also have a few *minor suggestions* for the README to make the instructions even clearer:
	 1. The instructions in the Manual Usage section cover creating a virtual environment, which is great.  It might also be useful to say earlier, in the Automatic Usage section, that the code will automatically create and activate a virtual environment called `venv` if one is not already available.
	 2. In [line 33]( https://github.com/enthusi/feline/blob/68243010ab2bc18a2740a4bcdcb2422fa988b95a/README.md?plain=1#L33)- I think you mean "into" not "in"
	 3. In all of the instructions in the Manual Usage section, it would be helpful to either explain what each of the flags means for each line, and what the options are; or to point to something like a sphinx documentation site where each option is explained.
	 4. It might be a good idea to state the size of the file that gets automatically downloaded and include instructions for doing that in the Manual Usage section as well.
 2. It would be good to have more documentation on the meaning of the results.  Axis labels are need on all of the plots.  A subtitle for the panel showing where the source was found would be good. What are the blue and grey lines in the panels of the spectra? What does match strength mean and what is a good value? What is b? What are the red, blue and orange lines in the emission line panels? What are the collapsed, white, quality, redshift and no lines panels? What do their colourmaps mean - can you put colour bars on them?
	 1. *Minor suggestion*: The second panel's title should probably be `14 lines used` not `14 used lines`
 3. There are quite a few files that are cleared up with the `make clean` command - it could be useful for the User to know what these are, so that they can make an informed decision about whether to delete them or not.  Somewhere in the documentation it would be helpful to give a list of all of these, explaining them.
 4. Can the User change the emission lines that are being looked for, for example if they were running the program on data from a different telescope/instrument?  Where could they change them if they wanted to?
 5. The Project Overview at the top of the README gives a nice description of what the code does.  I have a few questions that could do with some more information somewhere in the documentation: What galaxy emission line templates are fit at the end?  Can the User input their own options here? And how are numbers like the OII line ratio recorded?
 6. *Minor suggestion:* Towards the end of the Project Overview, this sentence: "Those extracted spectra are fitted with an emission line galaxy template and with the redshift as well as the individual line strengths as only free parameter to reach sub pixel accuracy in an early redshift estimate as well as deriving further diagnostics for the later manual inspection, such as the OII line ratio." - was a little confusing.  Consider splitting it into multiple sentences, to make it clearer.
 7. It's great that you have included a detailed statement on how people can contribute to the project, with instructions.  I did pick up, though, that in the CONTRIBUTING.md file you mention that documentation files are located in the `docs/` directory, but this doesn't exist.
 8. *Minor suggestion:* a little note somewhere in the documentation describing how the pdf is named (`results_date_time.fits`) would be nice - I only realised this after running it twice, and its a neat feature for keeping track of your result runs.
### Software Paper
1. The term "matched filter" is used a lot in the paper, without being defined.  I would suggest finding putting a sentence in describing what you mean.
2. The first figure needs colourbars
3. The second figure needs a caption, figure axis labels and colourbars.
4. *Minor Suggestion:* "Optionally, FELINE plots the three return parameters in real time via SDL surface along with storing them on disk." - somewhere, probably in your online documentation rather than in the paper, it would be good to say how to turn this option on and off.
5. Are there any other codes that do similar things to part of the code that you can compare to?  For example, the code fits the extracted 1D spectra for each galaxy candidate with an emission line template.  How does this compare to a ppxf fit of the spectra? Or a Bagpipes fit? Or do I have this wrong, and you're aiming at different results from these types of codes?
