### 2025-01-12 62bdedb
Update readme and header inclusion for Mac OS users.
