### 2025-01-12 enthusi/feline#10 @smutch
Please see #14 for a potential resolution to this. Mac OS users either need to install `gcc` themselves and use that as the compiler or install `libomp` and modify the compilation flags to get Clang to use that (e.g. `-fopenmp=libomp`).
