### 2025-01-13 d0382b5
update - add the to index.rst - add Makefile to build docs locally
