+++ docs/conf.py
@@ -1,0 +1,14 @@
+import sphinx_rtd_theme
+from docutils.nodes import author
+
+project = 'Feline'
+copyright = '2025, Feline Team'
+author = 'Martin Wendt, Marvin Henschel, Oskar Soth'
+release = '1.0.0'
+
+html_theme = 'sphinx_rtd_theme'
+
+html_theme_options = {
+	"repository_url": "https://github.com/enthusi/feline",
+	"use_repository_button": True,
+}
