+++ docs/index.rst
@@ -1,0 +1,5 @@
+Welcome to Feline (Find Emission Lines) Project's Documentation
+===============================================================
+
+Hello World
+-----------
