### 2025-01-13 979e06d
Update .readthedocs.yaml
added pip install theme
