+++ docs/requirements.txt
@@ -1,0 +1,1 @@
+spinx_rtd_theme
