### 2025-01-13 78c5066
changed version in docs/requirements.txt
