### 2025-01-13 e0fd6ec
added some autodoc rst files for testing purposes only
