+++ docs/src/Postprocessing/modules.rst
@@ -1,0 +1,9 @@
+Postprocessing
+==============
+
+.. toctree::
+   :maxdepth: 4
+
+   create_final_plots
+   create_pdf
+   detect_objects
