+++ docs/src/Preprocessing/modules.rst
@@ -1,0 +1,7 @@
+Preprocessing
+=============
+
+.. toctree::
+   :maxdepth: 4
+
+   combination
