+++ docs/src/Preprocessing/combination.rst
@@ -1,0 +1,7 @@
+combination
+==================
+
+.. automodule:: combination
+   :members:
+   :undoc-members:
+   :show-inheritance:
