+++ docs/src/Postprocessing/detect_objects.rst
@@ -1,0 +1,7 @@
+detect\_objects
+======================
+
+.. automodule:: detect_objects
+   :members:
+   :undoc-members:
+   :show-inheritance:
