### 2025-01-13 f892bee
minor bugfix in yaml file
