+++ .readthedocs.yaml
@@ -19,4 +19,5 @@ sphinx:
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
   install:
-    - requirements: docs/requirements.txt requirements.txt
+    - requirements: docs/requirements.txt
+                    requirements.txt
