+++ .readthedocs.yaml
@@ -20,4 +20,5 @@ sphinx:
 python:
   install:
     - requirements: docs/requirements.txt
-                    requirements.txt
+    - requirements: requirements.txt
+    - path: .
