+++ .readthedocs.yaml
@@ -21,4 +21,3 @@ python:
   install:
     - requirements: docs/requirements.txt
     - requirements: requirements.txt
-    - path: .
