### 2025-01-13 294efd7
minor bugfix in index.rst file
