### 2025-01-13 c8d985b
minor bugfix in index.rst file
