+++ docs/index.rst
@@ -4,9 +4,11 @@ Welcome to Feline (Find Emission Lines) Project's Documentation
 Hello World
 -----------
 Welcome to the Feline project! This documentation provides an overview of the project and its components.
+
+Contents
+--------
 .. toctree::
    :maxdepth: 2
-   :caption: Contents:
 
    src/Postprocessing/modules
    src/Preprocessing/modules
