### 2025-01-13 0bb5a61
fixed some linebreak issues in the autodoc code generation in sphinx
