+++ docs/index.rst
@@ -1,15 +1,40 @@
-Welcome to Feline (Find Emission Lines) Project's Documentation
-===============================================================
+|Github| |JOSS|
 
-Hello World
------------
+FELINE - A tool to detect emission line galaxies in 3d data
+===========================================================
+
+About
+-----
 Welcome to the Feline project! This documentation provides an overview of the project and its components.
 
-Contents
---------
+Contributing
+------------
+
 .. toctree::
    :maxdepth: 1
-   :caption: Code
+   :caption: Installation Guide
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Example Usage
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Documentation
 
    src/Postprocessing/modules
    src/Preprocessing/modules
+
+
+
+
+
+
+.. |Github| image:: https://img.shields.io/badge/GitHub-FELINE-blue
+   :target: https://github.com/enthusi/feline
+
+.. |JOSS| image:: https://joss.theoj.org/papers/a575acd1ffab0604de7e26eb83fd9bdc/status.svg
+   :target: https://joss.theoj.org/papers/a575acd1ffab0604de7e26eb83fd9bdc
+   :alt: status
+
+
