### 2025-01-14 72ca250
added installation.rst and added code copy button
