+++ docs/conf.py
@@ -13,6 +13,7 @@ extensions = [
 	'sphinx.ext.autodoc',
 	'sphinx.ext.napoleon',  # If you're using Google-style or NumPy-style docstrings
 	'sphinx.ext.viewcode',  # Optional: Adds links to highlighted source code
+    'sphinx_copybutton',
 ]
 
 project = 'Feline'
@@ -22,6 +23,7 @@ release = '1.0.0'
 
 html_theme = 'sphinx_rtd_theme'
 html_show_sourcelink = False
+pygments_style = 'sphinx'
 
 autodoc_default_options = {
     'members': True,
