+++ docs/usage/example_usage.rst
@@ -1,0 +1,3 @@
+Example Usage
+=============
+
