+++ docs/index.rst
@@ -9,6 +9,10 @@ Welcome to the Feline project! This documentation provides an overview of the pr
 
 Contributing
 ------------
+We welcome everyone who wants to contribute to our project you can find our
+Code of Conduct and Contribution Guidelines here:
+
+
 
 .. toctree::
    :maxdepth: 1
@@ -20,10 +24,15 @@ Contributing
 .. toctree::
    :maxdepth: 1
    :caption: Example Usage
+   :hidden:
+
+   usage/example_usage
+
 
 .. toctree::
    :maxdepth: 1
    :caption: Documentation
+   :hidden:
 
    src/postprocessing/modules
    src/preprocessing/modules
@@ -39,5 +48,3 @@ Contributing
 .. |JOSS| image:: https://joss.theoj.org/papers/a575acd1ffab0604de7e26eb83fd9bdc/status.svg
    :target: https://joss.theoj.org/papers/a575acd1ffab0604de7e26eb83fd9bdc
    :alt: status
-
-
