### 2025-01-14 5ab66f1
fix: - broken link - typo in feline instruction
update:
- add badge for docs
