+++ README.md
@@ -1,4 +1,4 @@
-[![License](https://img.shields.io/badge/License-BSD_3--Clause-blue.svg)](LICENSE) [![status](https://joss.theoj.org/papers/a575acd1ffab0604de7e26eb83fd9bdc/status.svg)](https://joss.theoj.org/papers/a575acd1ffab0604de7e26eb83fd9bdc)
+[![License](https://img.shields.io/badge/License-BSD_3--Clause-blue.svg)](LICENSE) [![status](https://joss.theoj.org/papers/a575acd1ffab0604de7e26eb83fd9bdc/status.svg)](https://joss.theoj.org/papers/a575acd1ffab0604de7e26eb83fd9bdc) [![Docs](http://img.shields.io/badge/Docs-latest-green.svg)](https://feline.readthedocs.io)
 # Feline (Find Emission Lines)
 ## Table of Contents
 - [Project Overview](#project-overview)
@@ -90,7 +90,7 @@ Now we can compile and run the main program
 ```bash
 cd ../../
 make
-./ feline.bin <ZLOW> <ZHIGH> <MAX_MATCH> <IGNORE_BELOW>
+./feline.bin <ZLOW> <ZHIGH> <MAX_MATCH> <IGNORE_BELOW>
 ```
 Now we start the postprocessing
 ```bash
@@ -111,6 +111,7 @@ data/pdf_files/result_*.pdf
 ```
 after running the workflow you can clean up all temporary files with
 ```bash
+cd ../..
 make clean
 ```
 
