### 2025-01-14 920dc5e
fixed some typos in installation instruction
