+++ Makefile
@@ -24,8 +24,11 @@ LDFLAGS += $(SDL2_LIBS)
 CFLAGS += -D SDLavailable=$(SDLavailable)
 
 
-CUBELINK := "https://amused.univ-lyon1.fr/data/UDF/HUDF/download/DATACUBE_UDF-10.fits"
-CUBENAME := "DATACUBE_UDF-10.fits"
+#CUBELINK := "https://amused.univ-lyon1.fr/data/UDF/HUDF/download/DATACUBE_UDF-10.fits"
+#CUBENAME := "DATACUBE_UDF-10.fits"
+
+CUBELINK := "martinwendt.de/cube.fits"
+CUBENAME := "cube.fits"
 
 ZLOW="0"
 ZHIGH="1.9"
