+++ src/feline.c
@@ -10,6 +10,7 @@
 #define MIN(x, y) (((x) < (y)) ? (x) : (y))
 #define DATA_PATH_PROCESSED "data/processed/"
 #define DATA_PATH_RUNTIME_FILES "data/runtime_files/"
+#define FELINE_VERSION "Feline v1.0.0"
 
 
 /**
@@ -105,8 +106,7 @@ int main(int argc, char *argv[]) {
 #endif
     int i, x, y;
 
-    printf("Feline v1.0.0 - 2024/09/12\n");
-    printf("Martin Wendt\n");
+    printf(FELINE_VERSION);
     if (argc < 5) {
         printf("Syntax: %s zlow zhigh max_match ignore_below\n", argv[0]);
         return 1;
@@ -194,7 +194,7 @@ int main(int argc, char *argv[]) {
 
     atexit(SDL_Quit);
 
-    window = SDL_CreateWindow("Feline v1.0.0 - Martin Wendt", SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED, dx * 4, dy + 1, SDL_WINDOW_SHOWN);
+    window = SDL_CreateWindow(FELINE_VERSION, SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED, dx * 4, dy + 1, SDL_WINDOW_SHOWN);
     if (window == NULL) {
         fputs(SDL_GetError(), stderr);
         exit(1);
