+++ src/postprocessing/project_path_config.py
@@ -28,4 +28,4 @@ DATA_PATH_PDF = os.path.join(DATA_PATH_ROOT, "data", "pdf_files")
 DATA_PATH_LOOKUP = os.path.join(DATA_PATH_ROOT, "data", "lookup")
 DATA_PATH_PROCESSED = os.path.join(DATA_PATH_ROOT, "data", "processed")
 DATA_PATH_RAW = os.path.join(DATA_PATH_ROOT, "data", "raw")
-DATA_PATH_RUNTIME_FILES = os.path.join(DATA_PATH_ROOT, "data", "runtime_fiels")
+DATA_PATH_RUNTIME_FILES = os.path.join(DATA_PATH_ROOT, "data", "runtime_files")
