+++ src/postprocessing/create_final_plots.py
@@ -333,7 +333,7 @@ if __name__ == "__main__":
 
     dz, dy, dx = cube.shape
 
-    catalog = open(sys.argv[3])
+    catalog = open(f"{project_path_config.DATA_PATH_RUNTIME_FILES}/{sys.argv[3]}")
     try:
         colors = mpl.colormaps['winter']
     except Exception as e:
