### 2025-01-16 9e5e120
update: - no files stored in root anymore - remove redundant variables
