+++ Makefile
@@ -24,24 +24,18 @@ LDFLAGS += $(SDL2_LIBS)
 CFLAGS += -D SDLavailable=$(SDLavailable)
 
 
-#CUBELINK := "https://amused.univ-lyon1.fr/data/UDF/HUDF/download/DATACUBE_UDF-10.fits"
-#CUBENAME := "DATACUBE_UDF-10.fits"
+CUBELINK := "https://amused.univ-lyon1.fr/data/UDF/HUDF/download/DATACUBE_UDF-10.fits"
+CUBENAME := "DATACUBE_UDF-10.fits"
 
-CUBELINK := "martinwendt.de/cube.fits"
-CUBENAME := "cube.fits"
 
 ZLOW="0"
 ZHIGH="1.9"
 MAX_MATCH="20"
 IGNORE_BELOW="7"
-# Define a variable to store the directory path
-MAKEFILE_DIR := $(dir $(abspath $(lastword $(MAKEFILE_LIST))))
-CUBEFILE := $(MAKEFILE_DIR)$(CUBENAME)
-CUBEFILENAME := $(CUBENAME)
 CORES := $(shell nproc)
 
 
-all: $(TARGET)                                   
+all: $(TARGET)
 
 $(TARGET): $(SOURCE)
 
@@ -50,11 +44,12 @@ $(TARGET): $(SOURCE)
 
 run:
 
-	@if [ -f $(CUBENAME) ]; then \
+	@if [ -f data/raw/$(CUBENAME) ]; then \
 		echo "File exists";\
 	else\
 		echo "Downloading Cube File...";\
 		wget --no-check-certificate $(CUBELINK) ;\
+		mv $(CUBENAME) data/raw/$(CUBENAME);\
 	fi
 
 	@if [ ! -d "venv" ]; then \
@@ -69,7 +64,7 @@ run:
 	@echo ""
 	@echo "apply median filter to 'flat' out the data (emission lines remain)..."
 	@. venv/bin/activate && \
-	python src/preprocessing/median-filter-cube.py $(CUBEFILE) --signalHDU=1 --varHDU=2 --num_cpu=$(CORES) --width=151 --output=data/processed/med_filt.fits > /dev/null
+	python src/preprocessing/median-filter-cube.py data/raw/$(CUBENAME) --signalHDU=1 --varHDU=2 --num_cpu=$(CORES) --width=151 --output=data/processed/med_filt.fits > /dev/null
 	@echo ""
 	@echo "filter the data cube with a 'typical line' template in spatial dimension first..."
 	@. venv/bin/activate && \
@@ -84,12 +79,11 @@ run:
 	@echo "deleting tmp files..."
 	@rm data/processed/spatial_cc.fits
 	@rm data/processed/spectral_cc.fits
-	@ln -s $(CUBEFILE) data/raw/
 
 	@echo "Create Masking Plot and transpose Cube for better Cache Access..."
 	@. venv/bin/activate ; \
 	cd src/preprocessing ; \
-	python combination.py $(CUBEFILENAME) s2n_v250.fits
+	python combination.py $(CUBENAME) s2n_v250.fits
 	@echo "Starting FELINE..."
 	@if [ -e feline.bin ]; then \
 		rm -f feline.bin; \
@@ -103,7 +97,7 @@ run:
 	@. venv/bin/activate ; \
 	cd src/postprocessing || exit ; \
 	python detect_objects.py s2n_v250.fits ; \
-	python create_final_plots.py $(CUBEFILENAME) s2n_v250.fits sorted_catalog.txt med_filt.fits J0014m0028 ; \
[…]
