### 2025-01-17 082f1fc
update: - add function to automatically insert current year to copyright in footer - remove unused line sin conf.py
