### 2025-01-17 a61211d
implemented a core fetching mechanism for macos
