### 2025-01-17 enthusi/feline#10 @osoth
We probably found a fix for fetching the cores with nproc on macOS. We changed the following line in the Makefile:

`CORES := $(shell uname -s| awk '{if ($$0 == "Darwin") print "sysctl -n hw.physicalcpu"; else print "nproc"}' | sh)
`
This should automatically detect the available cores on the system and store the value in CORES. However, we cannot verify it right now because we do not have a macOS system available.
