### 2025-01-17 c101574
update: - add note for mac os users for OMP
fix:
- typos in README.md
