### 2025-01-17 f4a5ac6
update: - remove unused setup.cfg
fix:
- branch in python-app.yml
