+++ .github/workflows/python-app.yml
@@ -2,9 +2,9 @@ name: Python application
 
 on:
   push:
-    branches: [ "dev" ]
+    branches: [ "main" ]
   pull_request:
-    branches: [ "dev" ]
+    branches: [ "main" ]
 
 permissions:
   contents: write
