+++ setup.cfg
@@ -1,3 +1,0 @@
-[tool:pytest]
-addopts = -v
-testpaths = test
