### 2025-01-17 77127b1
update: - add verbose mode for testing
