### 2025-01-17 8ac5098
update: - remove redundant code fragments
closes #20
