+++ src/postprocessing/create_final_plots.py
@@ -55,6 +55,9 @@ import ref_index
 import struct
 import project_path_config
 
+import detect_objects
+
+
 
 def scale_params(redshift: float) -> float:
     """
@@ -100,27 +103,6 @@ def get_impact(QSO_X: float, QSO_Y: float, px: float, py: float,
     scale = scale_params(z)
     # print theta,scale, theta*scale, b
     return theta * scale
-
-
-# MW23 it is ageneral annoyance to convert pixel positions in an
-# image to world coordinates (WCS) given as two angles on the sky (ra, dec)
-def pix_to_world(coord: astropy.wcs.WCS, pix: tuple) -> tuple:
-    """
-    Converts world coordinates (RA, Dec) to pixel coordinates.
-
-    Args:
-        coord (astropy.wcs.WCS): WCS object for coordinate transformation.
-        rad (tuple): World coordinates (RA, Dec) to convert.
-
-    Returns:
-        tuple: Pixel coordinates corresponding to the given world coordinates.
-    """
-
-    pixarray = np.array([[pix[0], pix[1], 0]], np.float64)
-    world = coord.wcs_pix2world(pixarray, 0)
-    ra = world[0][0]
-    dec = world[0][1]
-    return ra, dec
 
 
 # MW23 a single model ist just an integer number,
@@ -221,6 +203,32 @@ def galaxy(w: float, *p: int) -> np.ndarray:
     return flux
 
 
+def find_bottom_of_plot(ax, crval, crmax, data):
+    # find bottom:
+    lowest = min(data)
+    if lowest >= 0:
+        bottom = -10
+    if lowest < 0:
+        bottom = lowest * 1.2
+
+    ax.set_xticks(np.arange(crval, crmax, 200))
+    ax.set_xlim(crval, crmax)
+    ax.set_ylim(bottom, max(data) * 1.2)
+
+    return ax
+
+
+def add_ticks_to_plot(plt, aw, px_py):
+    plt.xlim(aw - 10, aw + 10)
+    plt.ylim(px_py - 10, px_py + 10)
+    plt.tick_params(axis="both", which="both", right=True,
+                    top=True, left=True, bottom=True,
+                    labelbottom=False, labelleft=False,
+                    labeltop=False, labelright=False)
+    plt.gca().invert_yaxis()
+
+    return plt
+
 if __name__ == "__main__":
     mpl.use("Agg")
     logging.disable(logging.CRITICAL - 2)
@@ -267,19 +275,11 @@ if __name__ == "__main__":
     columns = 12
     rows = 4
 
-    with open(os.path.join(project_path_config.DATA_PATH_PROCESSED,
-                           "raw_reordered_s2ncube.dat"), "rb") as f:
-        header = f.read()[:16]
+    dz, xd, yd, header = detect_objects.extract_arrays()
 
-    dz = struct.unpack("f", header[0:4])[0]
-    xd = struct.unpack("f", header[4:8])[0]
-    yd = struct.unpack("f", header[8:12])[0]
     crval = struct.unpack("f", header[12:16])[0]
     crmax = crval + dz * 1.25
 
-    dz = int(dz)
-    xd = int(xd)
-    yd = int(yd)
 
     if len(sys.argv) < 2:
         print("SYNTAX: %s cube.fits catalog.cat [ds9.reg]" % sys.argv[0])
@@ -297,14 +297,7 @@ if __name__ == "__main__":
                            "atoms.json"), "r") as data:
         atoms = json.load(data)
 
-    data = np.fromfile(os.path.join(project_path_config.DATA_PATH_RUNTIME_FILES,
-                                    "float32_array_omp4.raw"), dtype="float32")
[…]
