### 2025-01-19 enthusi/feline#10 @smutch
Ah - I never noticed the `nproc` issue as I have that command on my Mac from `coreutils` in homebrew.

I just tested your `CORES` variable defined above and it works perfectly on my Mac. 👍
