### 2025-01-21 bb35e02
update: - rearrange structure of gitignore - move benchmark script into benchmarl folder
