### 2025-01-21 cbca1a1
update: - add hash function for raw output of feline
