### 2025-01-21 9dcbb32
Makefile adjustements and requirements upper limit for scipy (#23)
* update:
- rearrange structure of gitignore
- move benchmark script into benchmarl folder

* update:
- add hash function for raw output of feline

* update:
- add "automatic" compiler detection
- add scipy pip version limit
