+++ requirements.txt
@@ -7,5 +7,5 @@ pypdf>=3.1.0
 mpdaf>=3.4
 matplotlib>=3.3.4
 astropy>=4.0
-scipy>=1.3.2
+scipy>=1.3.2,<1.14.0
 numpy>=1.17.4,<2.0.0
