### 2025-01-21 b08e3c7
update: - add clang as default compiler in Makefile - add notes for mac os and linux for how to use clang or gcc - reduce the content of the README.md and refer to the docs website
