### 2025-01-22 5df4a44
Modified Makefile and reduced Readme (#24)
* update:
- add clang as default compiler in Makefile
- add notes for mac os and linux for how to use clang or gcc
- reduce the content of the README.md and refer to the docs website

* fix:
- typos
