### 2025-01-22 3fc26a3
added user interaction for Download and environment
