+++ Makefile
@@ -26,7 +26,7 @@ CFLAGS += -D SDLavailable=$(SDLavailable)
 
 CUBELINK := "https://amused.univ-lyon1.fr/data/UDF/HUDF/download/DATACUBE_UDF-10.fits"
 CUBENAME := "DATACUBE_UDF-10.fits"
-
+CUBESIZE := $$(wget --spider --server-response --no-check-certificate $(CUBELINK) 2>&1 | awk -F '[()]' '/Length:/ {print $$2}' | tail -n 1)
 
 ZLOW="0"
 ZHIGH="1.9"
@@ -45,15 +45,29 @@ run:
 
 	@if [ -f data/raw/$(CUBENAME) ]; then \
 		echo "File exists";\
-	else\
-		echo "Downloading Cube File...";\
-		wget --no-check-certificate $(CUBELINK) ;\
-		mv $(CUBENAME) data/raw/$(CUBENAME);\
+	else \
+		read -p "Do you want to download the cubefile ($(CUBESIZE))? (y/n) " yn; \
+		if [ "$$yn" = "y" ]; then \
+			echo "Downloading Cube File..."; \
+			wget --no-check-certificate $(CUBELINK); \
+			mv $(CUBENAME) data/raw/$(CUBENAME); \
+		else \
+			echo "Download aborted. Exiting Makefile"; \
+			exit 1; \
+		fi \
 	fi
 
 	@if [ ! -d "venv" ]; then \
-		echo "Setting up environment..."; \
-		python3 -m venv venv; \
+		read -p "Do you want to create a Python virtual environment and install requirements? (y/n) " yn; \
+		if [ "$$yn" = "y" ]; then \
+			echo "Setting up environment..."; \
+			python3 -m venv venv; \
+		else \
+			echo "Exiting setup."; \
+			exit 1; \
+		fi \
+	else \
+		echo "Virtual environment already exists."; \
 	fi
 
 	@. venv/bin/activate; pip install --prefer-binary -r requirements.txt > /dev/null
@@ -103,15 +117,29 @@ cuda:
 
 	@if [ -f data/raw/$(CUBENAME) ]; then \
 		echo "File exists";\
-	else\
-		echo "Downloading Cube File...";\
-		wget --no-check-certificate $(CUBELINK) ;\
-		mv $(CUBENAME) data/raw/$(CUBENAME) ;\
+	else \
+		read -p "Do you want to download the cubefile ($(CUBESIZE))? (y/n) " yn; \
+		if [ "$$yn" = "y" ]; then \
+			echo "Downloading Cube File..."; \
+			wget --no-check-certificate $(CUBELINK); \
+			mv $(CUBENAME) data/raw/$(CUBENAME); \
+		else \
+			echo "Download aborted. Exiting Makefile"; \
+			exit 1; \
+		fi \
 	fi
 
 	@if [ ! -d "venv" ]; then \
-		echo "Setting up environment..."; \
-		python3 -m venv venv; \
+		read -p "Do you want to create a Python virtual environment and install requirements? (y/n) " yn; \
+		if [ "$$yn" = "y" ]; then \
+			echo "Setting up environment..."; \
+			python3 -m venv venv; \
+		else \
+			echo "Exiting setup."; \
+			exit 1; \
+		fi \
+	else \
+		echo "Virtual environment already exists."; \
 	fi
 
 	@. venv/bin/activate; pip install --prefer-binary -r requirements.txt > /dev/null
@@ -164,6 +192,3 @@ clean:
 	find data/processed ! -name '.gitkeep' -type f -delete
[…]
