### 2025-01-22 535f6a6
update: - make .raw file name more precise - remove more unused variables - add units to subplots on pdf
