### 2025-01-22 1ac4291
changed feline parallelization from dynamic to static because of same workload over all iterations. Changed Progress Bar accordingly
