### 2025-01-23 e5e079c
added compiler flags -march=native -mtune=native so it compiles feline directly for the native cpu. Leads to a small performance increase
