+++ Makefile
@@ -1,5 +1,5 @@
 CC = clang
-CFLAGS = -O3 -ffast-math -fopenmp=libomp -g -std=c99
+CFLAGS = -O3 -ffast-math -march=native -mtune=native -fopenmp=libomp -g -std=c99
 LDFLAGS = -lm
 TARGET = feline.bin
 SOURCE = src/feline.c
