+++ README.md
@@ -1,5 +1,4 @@
-[![status](https://joss.theoj.org/papers/a575acd1ffab0604de7e26eb83fd9bdc/status.svg)](https://joss.theoj.org/papers/a575acd1ffab0604de7e26eb83fd9bdc) [![Docs](http://img.shields.io/badge/Docs-latest-green.svg)](https://feline.readthedocs.io) [![License](https://img.shields.io/badge/License-BSD_3--Clause-blue.svg)](LICENSE)  [![Contributor Covenant](https://img.shields.io/badge/Contributor_Convenant-2.1-blue)](CONDUCT.md) [![CONTRIBUTION](https://img.shields.io/badge/CONTRIBUTING-1.0-blue)](CONTRIBUTING.md) 
-
+[![status](https://joss.theoj.org/papers/a575acd1ffab0604de7e26eb83fd9bdc/status.svg)](https://joss.theoj.org/papers/a575acd1ffab0604de7e26eb83fd9bdc) [![Docs](http://img.shields.io/badge/Docs-latest-green.svg)](https://feline.readthedocs.io) 
 
 
 # Feline (Find Emission Lines)
@@ -8,23 +7,24 @@
 
 Feline combines a fully parallelized galaxy line template matching with the matched filter approach for individual emission features of LSDcat. 
 For the 3D matched filtering, the complete data cube is first median filtered to remove all continuum sources, and then cross-correlated with a template of 
-an isolated emission feature in two spatial and one spectral dimension. We assumed a simple Gaussian with a FWHM of 250 km/s for the line profile and a PSF 
+an isolated emission feature in two spatial and one spectral dimension. We assumed a simple Gaussian with a FWHM of $250 \ \rm{km}/\rm{s}$ for the line profile and a PSF 
 based on the given seeing in the data cube. The FELINE algorithm then evaluates the likelihood in each spectrum of the cube for emission lines at the positions 
 provided by a given redshift and a certain combination of typical emission features. FELINE probes all possible combinations of up to 14 transitions paired 
 in 9 groups: $\rm{H}\alpha, \rm{H}\beta, \rm{H}\gamma, \rm{H}\delta$, $\rm{[OII]}$, $\rm{[OIII]}$, $\rm{[NII]}$, $\rm{[SII]}$, and $\rm{[NeIII]}$ for the redshift range of interest $(0.4 < z < 1.4)$. 
 This particular selection of lines is motivated by the most prominent emission features expected in the MUSE data within this redshift range. This results in $512 \ (2^9)$ 
-different models that are assessed at roughly 8,000 different redshifts for each of the approx 90,000 spectra in a single data cube. To ensure that only lines above 
-a certain S/N threshold contribute to each model, a penalty value is subtracted for each additional line. The S/N near strong sky lines are set exactly to that threshold. 
+different models that are assessed at roughly $8,000$ different redshifts for each of the approx $90,000$ spectra in a single data cube. To ensure that only lines above 
+a certain $\rm{S}\rm{N}$ threshold contribute to each model, a penalty value is subtracted for each additional line. The $\rm{S}/\rm{N}$ near strong sky lines are set exactly to that threshold. 
 Hence lines that fall onto such a contaminated region will not affect model quality. This is particularly useful for doublet lines that then contribute to a model even when one of 
-the lines aligns with a skyline. Furthermore, the S/N is saturated at a certain threshold to limit the impact of extremely strong lines on the overall budget 
+the lines aligns with a skyline. Furthermore, the $\rm{S}/\rm{N}$ is saturated at a certain threshold to limit the impact of extremely strong lines on the overall budget 
 of the tested template. For each spaxel the model with the highest accumulative probability over all contributing lines and its corresponding redshift are determined. 
[…]
