### 2025-01-27 5951777
update: - add missing doc strings
