### 2025-01-28 d596c71
update: - refactor python code for modularity and allow execution from the root directory without changing directories
