### 2025-01-28 520beb0
Update index.rst
adds an 'About' entry to the docs
