### 2025-01-28 c620445
update: - add autodoc to new create_final_plots.py functions
