+++ Makefile
@@ -25,8 +25,7 @@ CFLAGS += -D SDLavailable=$(SDLavailable)
 
 
 CUBELINK := "https://amused.univ-lyon1.fr/data/UDF/HUDF/download/DATACUBE_UDF-10.fits"
-#CUBENAME := "DATACUBE_UDF-10.fits"
-CUBENAME := "cube.fits"
+CUBENAME := "DATACUBE_UDF-10.fits"
 
 
 ZLOW="0"
