### 2025-01-28 275e3f6
Made python modular and consistent script calling from root directory (#26)
* update:
- rearrange structure of gitignore
- move benchmark script into benchmarl folder

* update:
- add hash function for raw output of feline

* update:
- add "automatic" compiler detection
- add scipy pip version limit

* update:
- make .raw file name more precise
- remove more unused variables
- add units to subplots on pdf

* update:
- refactor python code for modularity and allow execution from the root directory without changing directories

* update:
- typo fix
