+++ src/postprocessing/create_final_plots.py
@@ -53,10 +53,9 @@ import mpdaf
 import numpy as np
 import ref_index
 import struct
-import project_path_config
 
-import detect_objects
-
+import src.project_path_config as project_path_config
+import src.postprocessing.detect_objects as detect_objects
 
 
 def scale_params(redshift: float) -> float:
@@ -442,9 +441,15 @@ if __name__ == "__main__":
                   " y=%.1f, ra=%.6f dec=%.6f z=%.6f" %
                   (qso_id, run_id, px, py, ra, dec, z))
 
+        ax1.text(0.5, -0.2, r"wavelength ($\AA$)", ha="center", va="top",
+                 transform=ax1.transAxes, fontsize=10)
+
         ax2 = plt.subplot2grid((rows, columns), (1, 0), colspan=9)
         plt.title("%d used lines, match strength=%d, b=%.1f" % (
             used, quality, get_impact(qso_x, qso_y, px, py, z)))
+
+        ax2.text(0.5, -0.2, r"wavelength ($\AA$)", ha="center", va="top",
+                 transform=ax2.transAxes, fontsize=10)
 
         ax2.tick_params(
             axis="both",  # changes apply to the x-axis
