+++ src/preprocessing/project_path_config.py
@@ -1,7 +1,0 @@
-import os
-
-DATA_PATH_ROOT = os.path.abspath("../../")
-DATA_PATH_FINAL_PLOTS = os.path.join(DATA_PATH_ROOT, "data", "final_plots")
-DATA_PATH_LOOKUP = os.path.join(DATA_PATH_ROOT, "data", "lookup")
-DATA_PATH_PROCESSED = os.path.join(DATA_PATH_ROOT, "data", "processed")
-DATA_PATH_RAW = os.path.join(DATA_PATH_ROOT, "data", "raw")
