+++ src/postprocessing/detect_objects.py
@@ -44,7 +44,7 @@ import scipy.ndimage
 import skimage.feature
 import struct
 
-import project_path_config
+import src.project_path_config as project_path_config
 
 
 try:
@@ -166,7 +166,7 @@ def extract_arrays() -> tuple:
 def resize_filters(xd: int, yd: int) -> tuple:
     data = np.fromfile(
         os.path.join(project_path_config.DATA_PATH_RUNTIME_FILES,
-                     "float32_array_omp4.raw"), dtype="float32")
+                     "feline_float32_array.raw"), dtype="float32")
     plane, redshift, template, used = np.split(data, 4)
 
     plane.resize((xd, yd))
