### 2025-01-28 bc5cbc4
Updated Sphinx website (#27)
* update:
- add clang as default compiler in Makefile
- add notes for mac os and linux for how to use clang or gcc
- reduce the content of the README.md and refer to the docs website

* fix:
- typos

* fix:
- whitespace

* update:
- remove some badges

* update:
- change name back

* update:
- add missing doc strings

* update:
- add autodoc to new create_final_plots.py functions
