### 2025-01-28 3007ea2
first draft of about section
