+++ docs/usage/about.rst
@@ -1,0 +1,101 @@
+About
+============
+The FELINE algorithm then evaluates the likelihood in
+each spectrum of the 3D data cube for emission lines at 
+the positions provided by a given redshift and a certain
+combination of typical emission features.
+
+FELINE does not evaluate the observed data cube directly, but instead 
+utilizes the result of an emission line matched filter to boost the signal-to-noise
+of any such feature in the data cube.
+FELINE, however, does not pick individual peaks from that data as separate objects but instead
+simulatenously assesses the accumulative signal at all spectral positions that correspond to a certain
+set of emission lines at any redshift within the range of interest.
+
+FELINE probes all possible combinations of up to
+14 transitions paired in 9 groups:
+H$\alpha$, H$\beta$, H$\gamma$, H$\delta$, [O ii], [O iii], [N ii], [S ii], and Ne iii for the
+full redshift range of interest (e.g., 0.4 < z < 1.4).
+This particular selection of lines is motivated by the most prominent
+emission features expected in the MUSE data within the
+redshift range of [O ii] emitters.
+This results in 512 (2$^9$) different models that
+are assessed at roughly 8,000 different redshifts for each of
+the $\approx$ 90,000 spectra in a single data cube.
+
+
+Thresholds
+-------------
+To ensure that only lines above a certain S/N threshold 
+contribute to each model, a penalty value is subtracted for each additional
+line:
+
+.. code-block:: c
+   sum = sum - lines_ptr * scale * significance;
+
+The ``scale`` parameter (default = 10) enhances the signal of every considered line.
+``significance`` describes the minimum significance an additional line must have to result in 
+a net positive impact on the evaluation sum of the currently tested model (example: 7).
+This means that an additional feature below this significance has 0 effect on the final score.
+
+When running FELINE, the user also specifies a ``max_match`` parameter beyond which
+a line's additional significance is only added as logarithmic value (example: 20).
+This prevents extremely strong signals from dominating the parameter space.
+
+For each spaxel the model with the highest accumulative probability over all contributing
+lines and its corresponding redshift are determined. This approach has the benefit to pick
+up extremely weak emitters that show multiple emissions
+lines while avoiding the deluge of false positives when looking for single lines below a certain S/N threshold.
+This can be applied to each spatial element independently and was
+thus fully parallelized.
+
+Result
+-------------
+FELINE creates an binary file that contains the information that is being plotted during the evaluation
+(if an SDL environment is present).
+The created file contains raw 32bit data of four arrays - each with Y x X spatial dimension:
+
+- ``quality``, a float32 which reflects the score of the best matching template for this spaxel
+- ``redshift``, a float32 of the corresponding redshift of the best match
+- ``template``, an int32 the corresponding template of the best match
+- ``used``    , an int32 with the number of used line(pairs) in the template
+
+The latter simply representing a bit-count of the template integer value.
+
+Command line parameters
+-------------
+
+.. code-block:: bash
+
+      ./feline.bin <ZLOW> <ZHIGH> <MAX_MATCH> <IGNORE_BELOW>
+
+`ZLOW` and `ZHIGH` set the range of the redshift of interest. E.g., 0 - 1.9 (as in the example)
+
+`MAX_MATCH` is the threshold for very bright features. `20` yields reasonable results for all data cubes we
+handled. If you explicitly aim for weak objects, values in the range down to `10` can be tested. This parameter
+is not very critical.
+
+`IGNORE_BELOW` is the numerical value below which a feature in the matched filter signal-to-noise cube does not contribute
[…]
