### 2025-01-28 f51f382
patching the code blocks
