### 2025-01-28 21861b1
additions to docs
