+++ docs/src/postprocessing/detect_objects.rst
@@ -1,7 +1,0 @@
-detect\_objects
-======================
-
-.. automodule:: detect_objects
-   :members:
-   :undoc-members:
-   :show-inheritance:
