+++ docs/src/preprocessing/modules.rst
@@ -1,7 +1,0 @@
-Preprocessing
-=============
-
-.. toctree::
-   :maxdepth: 4
-
-   combination
