+++ docs/src/postprocessing/create_final_plots.rst
@@ -1,7 +1,0 @@
-create\_final\_plots
-===========================
-
-.. automodule:: create_final_plots
-   :members:
-   :undoc-members:
-   :show-inheritance:
