+++ docs/source/postprocessing/postprocessing.rst
@@ -1,0 +1,23 @@
+create\_final\_plots
+------------------------------------------
+
+.. automodule:: postprocessing.create_final_plots
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+create\_pdf
+---------------------------------
+
+.. automodule:: postprocessing.create_pdf
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+detect\_objects
+-------------------------------------
+
+.. automodule:: postprocessing.detect_objects
+   :members:
+   :undoc-members:
+   :show-inheritance:
