### 2025-01-28 33e8a9f
Changed sphinx autodoc and some typos and changed combination.py to masking_and_transpose.py (#29)
* update:
- rename combination.py to masking_and_transpose.py
- fix autodoc after filename change

* update:
- add missing files

* update:
- rename combination.py to masking_and_transpose.py
- fix autodoc after filename change

* update:
- add missing files

* update:
- add latex for rst format

---------

Co-authored-by: Marvin Henschel <marvin.henschel@uni-potsdam.de>
