+++ docs/src/preprocessing/combination.rst
@@ -1,7 +1,0 @@
-combination
-==================
-
-.. automodule:: combination
-   :members:
-   :undoc-members:
-   :show-inheritance:
