+++ docs/source/preprocessing/preprocessing.rst
@@ -1,0 +1,7 @@
+masking\_and\_transpose
+--------------------------------------------
+
+.. automodule:: preprocessing.masking_and_transpose
+   :members:
+   :undoc-members:
+   :show-inheritance:
