+++ docs/src/postprocessing/modules.rst
@@ -1,9 +1,0 @@
-Postprocessing
-==============
-
-.. toctree::
-   :maxdepth: 4
-
-   create_final_plots
-   create_pdf
-   detect_objects
