+++ docs/src/postprocessing/create_pdf.rst
@@ -1,7 +1,0 @@
-create\_pdf
-==================
-
-.. automodule:: create_pdf
-   :members:
-   :undoc-members:
-   :show-inheritance:
