### 2025-01-28 cebbd0b
update: - make docstrings less verbose
closes #21
