### 2025-01-28 c228d8d
update: - add note for disabling SDL with Makefile
