### 2025-01-28 c84fed1
see squashed commits (#30)
* update:
- rename combination.py to masking_and_transpose.py
- fix autodoc after filename change

* update:
- add missing files

* update:
- rename combination.py to masking_and_transpose.py
- fix autodoc after filename change

* update:
- add missing files

* update:
- add latex for rst format

* update:
- fix links representation

* update:
- make docstrings less verbose

closes #21

* update:
- add note for disabling SDL with Makefile

---------

Co-authored-by: Marvin Henschel <marvin.henschel@uni-potsdam.de>
