### 2025-01-29 #7528 @enthusi
@ivastar We are currently still addressing all of the valuable points made by the reviewers. The addition of a Sphinx based documentation turned out to be a very good idea but somewhat extensive. We are close to completion and  will prepare a report of the changes applied here.
