### 2025-01-29 b1d5aa2
fix: - remove italic units
