### 2025-01-29 2e9e430
added fixed window size for sdl and array truncation
