### 2025-02-03 7701c97
update: - move lsdcat code into a separate directory and add license and readme - update the lsdcat commands in docs - add PDF format to automatic usage
