+++ src/preprocessing/lsdcat/README.md
@@ -1,0 +1,1063 @@
+[//]: # (DO NOT EDIT THE FILE README.md IN THE ROOT OF THIS REPOSITORY)
+[//]: # (./doc/gendoc/README_tpl.md IS THE TEMPLATE THAT GENERATES THIS FILE)
+[//]: # (THE TEMPLATE GETS CONVERTED INTO ./README.md VIA  ./doc/gen_readme/gen_readme.sh)
+
+# LSDCat - *Line Source Detection and Cataloguing Tool* #
+
+LSDCat is a conceptually simple but robust and efficient detection
+package for emission lines in wide-field integral-field spectroscopic
+datacubes.  The detection utilises a 3D matched-filtering approach for
+compact single emission line objects.  Furthermore, the software
+measures fluxes and extents of detected lines.  LSDCat is implemented
+in Python, with a focus on fast processing of the large data-volumes
+of typical wide-field IFS datacubes.
+
+![LSDCat Logo](https://bitbucket.org/Knusper2000/lsdcat/raw/270da7489147414af85cc191e21a2226981fe6f8/doc/lsd_cat.jpg)
+
+**LSDCat2.0 introduced a major change to the algorithm - read the section
+"LSDCat2.0 - What's New" below to learn more.***
+
+The manual is work in progress.  While all the basics of LSDcat are
+covered, some of the more advanced features are not yet documented.
+Feedback, questions, or even pull requests are more than welcome!
+
+
+**Table of Contents**
+
+[TOC]
+
+## Requirements ##
+
+LSDCat runs on Python 3.  In order to use LSDCat you need the
+following 3rd party Python libraries installed on your system.
+
+  * astropy (>= 4) - http://www.astropy.org/
+  * NumPy (>= 1.18) - http://www.numpy.org/
+  * SciPy (>= 1.4) - http://www.scipy.org/
+
+A recommended python environment for astronomical research is [astroconda](https://astroconda.readthedocs.io/en/latest/).
+
+## Recommended Additional Software ##
+
+LSDCat can detect emission line objects, but (as of yet) not classify
+them.  This functionality is achieved by the companion software
+**QtClassify** (http://ascl.net/1703.011). 
+
+QtClassify can automatically identify galaxies that have multiple
+emission lines detected.  It also offers an easy-to-use graphical
+interface that allows for the interactive classification of single
+line detections.  For more information, installation-, and usage
+instructions visit the QtClassify web page:
+https://bitbucket.org/Leviosa/qtclassify
+
+
+## Installing LSDCat ##
+
+In your `/home/ehere` (or wherever you want to have LSDCat installed) do:
+
+`git clone https://bitbucket.org/Knusper2000/lsdcat.git`
+
+Then set up your `PYTHONPATH` to contain the LSDCat library directory
+`lsdcat/lib`.  For example if you have installed LSDCat in `/home/ehere` and
+use bash put in your `.bashrc`
+
+`export PYTHONPATH=${PATH}:${HOME}/lsdcat/lib/`
+
+To have the executables available anywhere on your platform, configure
+the systems `$PATH` variable accordingly.  Again, assuming you have
+installed LSDCat in your `$HOME` and you are using `bash` add to your
+`.bashrc`:
+
+`export PATH=${PATH}:${HOME}/lsdcat/`
+
+If you want to use the additional tools that are shipped with LSDCat
+(see below), put
+
+`export PATH=${PATH}:${HOME}/lsdcat/tools/`
+
+in your `.bashrc`.
+
+Users of other shells (e.g., `csh` or `tcsh`) have to follow [a
+different procedure](https://kb.iu.edu/d/acar). 
+
+## License ##
+
+LSDCat is licensed under a
[…]
