### 2025-02-03 24370c5
adding model fit and restricted catalog size
