### 2025-02-03 41ac85b
added hash test for example data cube
