### 2025-02-03 5011b33
added integrity test for the example data cube
