### 2025-02-03 ecc921c
function doc strings
