### 2025-02-03 c59bc06
example plot in docs
