+++ src/postprocessing/detect_objects.py
@@ -245,7 +245,7 @@ if __name__ == "__main__":
     # data=plane
     coordinates = skimage.feature.peak_local_max(
         data, min_distance=1, threshold_abs=50,
-        exclude_border=10, num_peaks=100)
+        exclude_border=10, num_peaks=50)
 
     width = 0
     bleft = width
