### 2025-02-03 ef590a2
add figure link and proper references in About
