+++ docs/usage/example_vis.rst
@@ -3,7 +3,7 @@ Example visualization
 We provide a Python script to visualize the FELINE output.
 The -> Automatic usage executes it as well.
 
-The FELINE output contains a raw 32bit data of four arrays - each with X:math:`\times`Y spatial dimension:
+The FELINE output contains a raw 32bit data of four arrays - each with X :math:`\times` Y spatial dimension:
 - ``quality``, a float32 which reflects the score of the best matching template for this spaxel
 - ``redshift``, a float32 of the corresponding redshift of the best match
 - ``template``, an int32 the corresponding template of the best match
@@ -12,7 +12,7 @@ The FELINE output contains a raw 32bit data of four arrays - each with X:math:`\
 The plot script creates a multipage PDF document with one page per potential galaxy for quick
 manual checking and verification ordered by quality of the detection:
 
-.. image:: path/exemplary_plot.png
+.. image:: exemplary_plot.png
   :alt: example of a plot for one galaxy with low continuum but strong emission
 
 In the the top right the plot shows the full quality map of the field of view with
