+++ docs/usage/automatic_usage.rst
@@ -1,3 +1,4 @@
+.. _automaticusage-label:
 Automatic Usage
 ---------------
 To use FELINE in an automated manner, you can use the pre-configured Makefile workflow.
