+++ docs/usage/example_vis.rst
@@ -1,7 +1,7 @@
-Example visualization
+Plots of individual detections
 ======
 We provide a Python script to visualize the FELINE output.
-The :ref:`_automaticusage-label` executes it as well.
+The :ref:`automaticusage-label` executes it as well.
 
 The FELINE output contains a raw 32bit data of four arrays - each with X :math:`\times` Y spatial dimension:
 - ``quality``, a float32 which reflects the score of the best matching template for this spaxel
