### 2025-02-03 12c1afa
adjust redshift grid step size
