### 2025-02-03 2369674
update: - add rounded raw binary array hash test
