### 2025-02-04 1226382
update: - remove truncate function from c code - add round raw array test
