### 2025-02-04 1b758c5
lsdcat dir in Makefile and testing published cube!
