### 2025-02-04 321cae5
update: - add hash from cuda - modify test part of the docs - rename test file
fix:
- typos in Makefile Cuda target
- typos in .rst files
