+++ docs/usage/example_vis.rst
@@ -1,5 +1,5 @@
 Plots of individual detections
-======
+==============================
 We provide a Python script to visualize the FELINE output.
 The :ref:`automaticusage-label` executes it as well.
 
