+++ docs/usage/about.rst
@@ -33,8 +33,8 @@ FELINE was utilized in the following publications:
 | - `MusE GAs FLOw and Wind (MEGAFLOW) IX. The impact of gas flows on the relations between the mass, star formation rate, and metallicity of galaxies <https://ui.adsabs.harvard.edu/abs/2023MNRAS.521..546L>`_
 
 
-.. image:: megaflow_survery.png
-  :alt: Detection histogram from the MEGAFLOW XII paper.
+.. image:: megaflow_survey.png
+   :alt: Detection histogram from the MEGAFLOW XII paper.
   
 This histogram from the `MEGAFLOW XII paper (Fig. 15) <https://ui.adsabs.harvard.edu/abs/2024arXiv241107014B>`_ reflects
 the typical usecase of FELINE.
@@ -83,7 +83,8 @@ The created file contains raw 32bit data of four arrays - each with Y x X spatia
 The latter simply representing a bit-count of the template integer value.
 
 .. image:: feline_result.png
-  :alt: Bitmaps of the FELINE result.
+   :alt: Bitmaps of the FELINE result.
+    
   
 Command line parameters
 -----------------------
