+++ docs/usage/manual_usage.rst
@@ -100,6 +100,6 @@ Alternatively, you can manually execute each step of the FELINE workflow. Follow
       make clean
       
 .. image:: running_on_80cores.png
-  :alt: SDL Window of Feline running on a 80 cores machine.
+   :alt: SDL Window of Feline running on a 80 cores machine.
   
 SDL Window of Feline running the UDF MUSE field on a 80 cores machine.
