+++ docs/usage/about.rst
@@ -38,7 +38,7 @@ FELINE was utilized in the following publications:
   
 This histogram from the `MEGAFLOW XII paper (Fig. 15) <https://ui.adsabs.harvard.edu/abs/2024arXiv241107014B>`_ reflects
 the typical usecase of FELINE.
-It shows that there is a significant popultion of galaxies detected solely on their emission line with FELINE (no continuum).
+It shows that there is a significant population of galaxies detected solely based on their emission line with FELINE (no continuum).
   
 For further Information see the `JOSS Paper <https://joss.theoj.org/papers/a575acd1ffab0604de7e26eb83fd9bdc>`_.
 
@@ -84,7 +84,7 @@ The latter simply representing a bit-count of the template integer value.
 
 .. image:: feline_result.png
    :alt: Bitmaps of the FELINE result.
-    
+           
   
 Command line parameters
 -----------------------
