### 2025-02-04 2db1faa
update: - add git clone instruction
fix:
- add typehint syntax for version below 3.10
