### 2025-02-04 #7528 @enthusi
Dear @Smutch, @bronreichardtchu, @ivastar
thanks alot for your helpful review.
We addressed all of your comments.
Namely, and most importantly, we added Sphinx documentation
attached to the project.

Regarding the comments by @smutch
1.1
 The codebase is restructured now.
 We moved LSDcat from `src/preprocessing` to `src/preprocessing/lsdcat`
 to separate it from our codebase.
 Additionally, we added the LSDcat license and README to `src/preprocessing/lsdcat`
 to clearly distinguish it from our own code.
 Furthermore, we refactored all duplicated code fragments of our code.
1.2
 The About section in the docs describes and links to LSDcat
 We decided to include the relevant parts of LSDcat into our project for three reasons:
 - the code needed to be cleaned to conform to pep8 standard
 - the code has changed its interface drastically in the past (different indices and alike)
 - the BSD licence made this feasible
1.3
 We made a series of changes to allow compiling under all three main OSes as we intended it to
 be right from the start but with the lack of a MacOS system for testing.
 Thanks to your commits and feedback this should be fully compatible now.
 In the course of this change, we moved to clang/llvm as main default compiler which
 simplifies the process for all potential users in our view.
 It is confirmed to compile and run via WSL 2 under Windows and even on Linux RiscV architecture now.
2.1
 The different pieces of source, documentation and output as well as temporary files
 are now properly placed in separate locations.
3.1
 The About part now describes the main FELINE output (the binary arrays).
3.2
 The docstrings were adjusted and streamlined wherever necessary
3.3
 This is addressed in the Sphinx documentation now - we further emphasized the point
 that the visualization of the FELINE output is in exemplary additional post processing
 step.
3.4
 We removed the CI tests completely.
3.5
 We added a new verification test for the raw FELINE output binary array.
 In this test, we compare the hash of the array from the C and CUDA results.
 Due to differences in single-precision floating-point operations between C and CUDA,
 we rounded the arrays to \(10^{-1}\) which is well within the required precision to
 confirm identicalness.
 A further optional test verifies the hash of the galaxy detection table extracted from that binary data.
 We also included more information on how to run the test in the Sphinx docs.
4.1
 A statement was added to the paper.
 While specialized software focuses on detecting individual emission features,
 traditional galaxy identification tools analyze entire images and often miss
 objects with minimal continuum flux (most notably, Sextractor by Bertin & Arnouts 1996).
 In contrast, the ``FELINE`` tool has been successfully tested as an automated identification system,
 providing rapid results for 3D data cubes that can be easily reviewed by humans.
 We added a histogram plot from the paper to the About section.

Regarding the comments by @bronreichardtchu
General Checks
1.
 Fully addressed and fixed.
2.
 Thank you for your valuable feedback!
 We’ve addressed all the issues with the help you pointed out.
 We added a more general solution for handling nproc on macOS and updated the Makefile.
 We added a note in the docs on how to switch to gcc when needed.
3.
 We were not able to reproduce this problem. We tried to identify it, but our research
 remained inconclusive. Instead we were asked to refrain from using Python versions that are
 stable but still in the bug fix phase as those are prone to changes.
 Currently Python 3.11 is the highest supported version, which is why we chose it as criterion.

Functionality
1.
 The web interface of github directly offers that service, but
 we added a instruction in the Readme and in the docs.

Documentation
1.
 We agree and implemented all mentioned points.
 1.4 Automatic usage:  If users run the Makefile
 they get asked if he/she wants to download the cube file (with cube size) and if they want to create the virtual environment.
2.
 The caption was changed.
3.
 We added a directory `data/runtime_files` which stores all files during the run .
 Make clean removes them but the user has the chance to copy them before removing them completely.
4.
 Yes, we stored the line pairs in a dedicated file and briefly explain the spots to change
 in the https://feline.readthedocs.io/en/latest/usage/about.html#modifications
 It contains all lines of interest for the intended usage. Modifications for a
 significant higher number of lines, etc, would require further changes.
 I am happy to support requests here to some extent.
5.
 We made a clearer distinction between FELINE output and possible ways for visualization
 and further analysis. The quick model fitting is merely for a good guess on the redshift.
 In the actual science case, we tried to use the OII line ratio as an automatic parameter
 to distinguish between OII and Ly-alpha, but it turns out to require human verification.
6.
 The paper text was changed in that regard.
7.
 Fixed.
8.
 We improved that aspect now.

Software Paper
1.
 We added a better introduction now.
2.
 The colors in all figures represent mostly arbitrary units and are solely used as contrast
 to differentiate separate sources. We found individual colorbars and labels to be too confusing and of little use.
 The caption refers to that now.
3.
 See 2.
4.
 We made related changes to the code and documentation.
5.
 See comment to 4.1 above.
 In addition: the extracted 1D spectra are just one possible way to visualize the FELINE detections.
 The actual detection of whole galaxies, even with low or no continuum flux is key for FELINE.
 line ratios and fluxes as well as optimally PSF extracted and mask-segmented spectra are a job best done
 with specialized software.
 (see also the references papers that use FELINE).

To both:
We are very grateful for all the useful comments that helped us improving the presentation of the software
and in particular the file structure and documentation.

We hope that our additions and adjustments fulfill your expectations of the required polishing for publication.
Sincerely,

Martin, Marvin & Oskar
