### 2025-02-10 #7528 @smutch
@enthusi - This is looking great! The code base looks a lot cleaner now, the documentation site is good, and it's great to see the regression tests added.

The new change to using clang in the Makefile didn't work for me (which maybe partially my fault!). It seems that using openmp with Clang on MacOS is a moving target! I will open a separate PR/issue for that.

I think the last thing that is required for me to complete my reviewer checklist is a discussion on the "state of the field" in the software paper (describing how this software compares to other commonly-used packages). Other than that, I think it looks great and, as soon as this is addressed, I'm happy to sign off.

Congrats on a nice package that I'm sure will be a great resource for the community!
