### 2025-02-10 enthusi/feline#31 @smutch
Clang and OpenMP on MacOS
It seems that with more recent versions of Mac OS, the openmp can't be used with the system-provided Clang, even if libomp is installed. Mac OS users will need to further install the LLVM toolchain to get a version of Clang that works with openmp:

```
brew install llvm
```

In order to not interfere with the system-provided Clang, homebrew will not put the new clang executable in the users PATH. Therefore the path to the homebrewed clang will need to be added to the Makefile:

```Makefile
CC = $(shell brew --prefix llvm)/bin/clang
```
