### 2025-02-10 #7528 @smutch
See https://github.com/enthusi/feline/issues/31 for the new Mac OS build issue.
