### 2025-02-10 enthusi/feline#31 @osoth
Thank you for the fast reply to our changes. I think we could maybe fix this by adding the following `if`-command to the Makefile:
```make
CC = clang
CFLAGS = -O3 -ffast-math -march=native -mtune=native -fopenmp=libomp -g -std=c23
LDFLAGS = -lm
TARGET = feline.bin
SOURCE = src/feline.c

ifeq ($(shell uname -s), Darwin)
    ifeq ($(CC), clang)
        CC := $(shell brew --prefix llvm)/bin/clang
    endif
endif
```
Then, we would only need to update the notes in the README and documentation to inform users that they need to install LLVM Clang to use OpenMP. Everything else can remain the same. This might be the best solution because nothing would change for Linux and Windows WSL users.
