+++ joss/paper.md
@@ -44,7 +44,14 @@ asks for automated processes to support the scientists.
 ￼
 The ``FELINE`` algorithm evaluates the likelihood of emission lines at specific positions in each spectrum of the data cube. It does this by probing all possible combinations of up to 14 typical emission features, including Hα, Hβ, Hγ, Hδ, [OII], [OIII], [NII], [SII], and [NeIII], for the redshift range of interest (0.4 < z < 1.4). This extensive analysis leads to approximately 230,000,000,000  iterations.
 
-While specialized software focuses on detecting individual emission features, traditional galaxy identification tools analyze entire images and often miss objects with minimal continuum flux. In contrast, the ``FELINE`` tool has been successfully tested as an automated identification system, providing rapid results for 3D data cubes that can be easily reviewed by humans.
+There are generally two approaches to source identification. Image based, such as SExtractor ([@Bertin1996]) and its many
+derivatives or emission line based, i.e., the ORIGIN code ([@MaryMary2020]) or the aforementioned LSDcat.
+The former group analyzes white or narrow band images and is prone to miss objects with minimal or no continuum flux.
+While specialized software that focuses on detecting emission features is generally aiming for individual lines, like Lyman-$\alpha$. In contrast, the ``FELINE`` tool evaluates a physically connected series of emission lines simultaneously
+and has been successfully tested as an automated identification system, providing rapid results for 3D data cubes that can be easily reviewed by humans.
+In [@Bouche2024] we use a dual galaxy identiﬁcation process based both on continuum (using SExtractor) and on
+emission lines using FELINE. Overall, we find that up to one third of our detected galaxies have no continuum
+detection within our magnitude limit.
 
 # Science field
 The signal-to-noise cube generated after matched filtering with a 3D emission line template reflects the probability of an emission line at a given spatial and spectral position,
