+++ joss/paper.bib
@@ -158,3 +158,20 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Mary2020,
+       author = {{Mary}, David and {Bacon}, Roland and {Conseil}, Simon and {Piqueras}, Laure and {Schutz}, Antony},
+        title = "{ORIGIN: Blind detection of faint emission line galaxies in MUSE datacubes}",
+      journal = {\aap},
+     keywords = {methods: data analysis, techniques: imaging spectroscopy, galaxies: high-redshift, methods: statistical, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2020,
+        month = mar,
+       volume = {635},
+          eid = {A194},
+        pages = {A194},
+          doi = {10.1051/0004-6361/201937001},
+archivePrefix = {arXiv},
+       eprint = {2002.00214},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020A&A...635A.194M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
