+++ joss/paper.bib
@@ -175,3 +175,17 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2020A&A...635A.194M},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+        
+@ARTICLE{Bertin1996,
+       author = {{Bertin}, E. and {Arnouts}, S.},
+        title = "{SExtractor: Software for source extraction.}",
+      journal = {\aaps},
+     keywords = {METHODS: DATA ANALYSIS, TECHNIQUES: IMAGE PROCESSING, GALAXIES: PHOTOMETRY},
+         year = 1996,
+        month = jun,
+       volume = {117},
+        pages = {393-404},
+          doi = {10.1051/aas:1996164},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1996A&AS..117..393B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
