### 2025-02-11 #7528 @enthusi
@smutch , @ivastar
Thank you for your quick reply again.
We improved the part that describes other software in the  "state of the field" of the paper and added some more explanation to the figure shown in the about section in the docs now:
https://feline.readthedocs.io/en/latest/usage/about.html
