### 2025-02-12 7a97b71
Added native clang to Makefile and Documentatioin
