+++ Makefile
@@ -1,9 +1,14 @@
 CC = clang
-CFLAGS = -O3 -ffast-math -march=native -mtune=native -fopenmp=libomp -g -std=c99
+CFLAGS = -O3 -ffast-math -fopenmp=libomp -g -std=c99
 LDFLAGS = -lm
 TARGET = feline.bin
 SOURCE = src/feline.c
 
+ifeq ($(shell uname -s), Darwin)
+    ifeq ($(CC), clang)
+        CC := $(shell brew --prefix llvm)/bin/clang
+    endif
+endif
 
 # Check for SDL2 availability
 SDL_CONFIG := sdl2-config
