### 2025-02-12 enthusi/feline#31 @osoth
[Commit 7a97b71](https://github.com/enthusi/feline/commit/7a97b712f9bdd8f540a1eed5646621c2a94a9e8a) closes #31
