### 2025-02-13 #7528 @bronreichardtchu
Everything looks good to me - the code runs well for me now, and the paper describes much more clearly where FELINE sits in comparison to other packages.  I also like the documentation site, it makes it much easier to understand what to use the code for.

I'm happy to sign off on this as well. Congrats on a great software package!
