### 2025-03-04 #7528 @ivastar
@enthusi Below are some editorial comments. Please review and implement (as long as you agree).

Title: Please capitalize as a title(e.g., A Tool to Detect Emission...)
Line 10: broader --> larger
Line 11: flux-level -> flux level
Line 13: producing --> produce
Line 20: in THE spectral direction
Line 20: maximiZes
Line 21: and still yields --> while still producing
Line 23: medium resolution --> medium-resolution
Line 25: amount --> volume
Line 25: asks for --> requires
Line 30: Maybe use scientific notation? 2.3x10^11
Line 32: i.e. --> such as
Line 33: aforementioned --> previously mentioned, doesn't hurt to add the citation again
Line 34: While, --> On the other hand,
Line 35: aiming for --> targeted at
Line 36: series --> set
Line 46: remove "substantially" since there the paper does not quantify the increase in significance or cite a figure if one is available
Line 47-48: no or little --> little or no
Line 53: through --> of
Line 58: at which --> where
Line 23, 70: R. Bacon, Roland Bacon --> Bacon (all other citations are last name only)
Line 71, 76: re-arranging --> rearranging
Line 72: amount --> number
Line 74: which would be strongly - > that would be highly
Line 75: spectrum exceeding -- > spectrum would exceed, this last sentence in the paragraph maybe should be edited without the "would"s, it's a bit hard to parse
Line 77: That --> This
Line 79: CUDA capable --> CUDA-enabled
Line 81: three return parameters - which ones, please specify? maybe I missed it.
Also line 81: is this a reference to SDL_Surface (in which case there should be an underscore)
Also line 81: along with storing them on disk --> and also saves them to disk
Line 83: left to right ARE the quality
Line 87: possible edit of this sentence: We also provide a simple Python framework for further visualization and manual verification of the FELINE detections (see Figure 2).
Line 83: I am not sure the figure this paragraph refers to is actually included in the text?
Figure 1: Did not notice reference in text, please add.

If there are any major packages that FELINE has a substantial dependency, I would recommend adding citations to those in the acknowledgements.

bib file:
Please replace \aap, \mnras and \aaps with the names of the journals, our system cannot parse them automatically.
Please replace the Bouche 2024 reference with the A&A reference.
