### 2025-03-04 #7528 @ivastar
> - No DOI given, and none found for title: MusE GAs FLOw and Wind (MEGAFLOW) XII. Rationale a...

There are a few MEGAFLOW paper in the *.bib file that are not actually cited in the paper. I would recommend removing them. The whole bib file is parsed even if not all citations are in the text and this seems to be raising an error from the DOI checker because the reference doesn't list it.
