### 2025-03-05 c50debe
Made first changes to the paper according to the checklist in #32
