+++ joss/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'FELINE: A tool to detect emission line galaxies in 3d data'
+title: 'FELINE: A Tool to Detect Emission Line Galaxies in 3D Data'
 tags:
   - Python
   - astronomy
@@ -31,23 +31,23 @@ aas-journal: Astrophysical Journal <- The name of the AAS journal.
 
 
 # Summary
-The detection and classification of objects in astrophysical data has been a key task since the earliest days of astronomy. Over the past decade, the volume of newly observed data has increased dramatically. The advent of integral field unit spectrographs (IFUs), which produce 3D data cubes, has shifted the focus from classical single-target observations to much broader fields of view captured in a single exposure.
-Simple flux-level peak detection algorithms based on thresholding are prone to either missing many potential real objects or, as a trade-off, producing an abundance of false positives.
+The detection and classification of objects in astrophysical data has been a key task since the earliest days of astronomy. Over the past decade, the volume of newly observed data has increased dramatically. The advent of integral field unit spectrographs (IFUs), which produce 3D data cubes, has shifted the focus from classical single-target observations to much larger fields of view captured in a single exposure.
+Simple flux level peak detection algorithms based on thresholding are prone to either missing many potential real objects or, as a trade-off, produce an abundance of false positives.
 
 # Statement of need
 The *F*ind *E*mission *LINE*s tool ``FELINE`` combines a fully parallelized galaxy line template matching with the matched filter approach for individual emission features of LSDcat [@HerenzE_17a; @herenz2023].
-This transformation cross-correlates the dataset with a template that ideally matches the properties of the signal to be detected. Here, the images are cross-correlated with a 2D template for the expected light distribution of the sources to be detected and the same is applied in spectral direction. The filter maximises the signal-to-noise ratio (S/N) of a source that is optimally represented by the template and still yields a significant S/N enhancement for similar signals.
+This transformation cross-correlates the dataset with a template that ideally matches the properties of the signal to be detected. Here, the images are cross-correlated with a 2D template for the expected light distribution of the sources to be detected and the same is applied in the spectral direction. The filter maximizes the signal-to-noise ratio (S/N) of a source that is optimally represented by the template while still producing a significant S/N enhancement for similar signals.
 
-The VLT/MUSE [@Bacon+10; @Bacon+14] 3D spectrograph creates $\sim$ 90,000 medium resolution spectra arranged in a 300 $\times$ 300 spatial grid.
-These data cubes have typical sizes of 3-6 GiB per exposure, the sheer amount of data
-asks for automated processes to support the scientists.
+The VLT/MUSE [@Bacon+10; @Bacon+14] 3D spectrograph creates $\sim$ 90,000 medium-resolution spectra arranged in a 300 $\times$ 300 spatial grid.
+These data cubes have typical sizes of 3-6 GiB per exposure, the sheer volume of data
+requires automated processes to support the scientists.
 ￼
-The ``FELINE`` algorithm evaluates the likelihood of emission lines at specific positions in each spectrum of the data cube. It does this by probing all possible combinations of up to 14 typical emission features, including Hα, Hβ, Hγ, Hδ, [OII], [OIII], [NII], [SII], and [NeIII], for the redshift range of interest (0.4 < z < 1.4). This extensive analysis leads to approximately 230,000,000,000  iterations.
[…]
