### 2025-03-05 3ac4085
edited the Bacon citation error because of inconsitent naming
