+++ joss/paper.bib
@@ -39,7 +39,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Bacon+22,
-       author = {{Bacon}, Roland and {Brinchmann}, Jarle and {Conseil}, Simon and {Maseda}, Michael and {Nanayakkara}, Themiya and {Wendt}, Martin and {Bacher}, Raphael and {Mary}, David and {Weilbacher}, Peter M. and {Krajnovi{\'c}}, Davor and {Boogaard}, Leindert and {Bouch{\'e}}, Nicolas and {Contini}, Thierry and {Epinat}, Beno{\^\i}t and {Feltre}, Anna and {Guo}, Yucheng and {Herenz}, Christian and {Kollatschny}, Wolfram and {Kusakabe}, Haruka and {Leclercq}, Floriane and {Michel-Dansac}, L{\'e}o and {Pello}, Roser and {Richard}, Johan and {Roth}, Martin and {Salvignol}, Gregory and {Schaye}, Joop and {Steinmetz}, Matthias and {Tresse}, Laurence and {Urrutia}, Tanya and {Verhamme}, Anne and {Vitte}, Eloise and {Wisotzki}, Lutz and {Zoutendijk}, Sebastiaan L.},
+       author = {{Bacon}, R. and {Brinchmann}, Jarle and {Conseil}, Simon and {Maseda}, Michael and {Nanayakkara}, Themiya and {Wendt}, Martin and {Bacher}, Raphael and {Mary}, David and {Weilbacher}, Peter M. and {Krajnovi{\'c}}, Davor and {Boogaard}, Leindert and {Bouch{\'e}}, Nicolas and {Contini}, Thierry and {Epinat}, Beno{\^\i}t and {Feltre}, Anna and {Guo}, Yucheng and {Herenz}, Christian and {Kollatschny}, Wolfram and {Kusakabe}, Haruka and {Leclercq}, Floriane and {Michel-Dansac}, L{\'e}o and {Pello}, Roser and {Richard}, Johan and {Roth}, Martin and {Salvignol}, Gregory and {Schaye}, Joop and {Steinmetz}, Matthias and {Tresse}, Laurence and {Urrutia}, Tanya and {Verhamme}, Anne and {Vitte}, Eloise and {Wisotzki}, Lutz and {Zoutendijk}, Sebastiaan L.},
         title = "{The MUSE Hubble Ultra Deep Field surveys: Data release II}",
       journal = {\aap},
      keywords = {galaxies: high-redshift, galaxies: distances and redshifts, techniques: imaging spectroscopy, catalogs, Astrophysics - Astrophysics of Galaxies},
