### 2025-03-05 c29c76e
update: - figure 1 reference fix: - modify bibtex file
