+++ joss/paper.md
@@ -94,7 +94,7 @@ Optionally, FELINE plots the three return parameters in real time via SDL_Surfac
 
 ![The visual FELINE output displays four metrics: signal strength, template, redshift, and line count. These values are presented using a shared color scheme to distinguish between the various peaks and objects. \label{fig:results}](feline_result.png)
 
-Shown are from left to right are the quality of the best match, the corresponding redshift of the best match and its template. A fourth panel shows the number of lines that contributed to the most successful model for ease of human readability (it reflects the number of set bits in the best model value).
+Shown are from left to right are the quality of the best match, the corresponding redshift of the best match and its template. A fourth panel shows the number of lines that contributed to the most successful model for ease of human readability (it reflects the number of set bits in the best model value). (See Figure 2)
 
 We also provide a simple Python framework for further visualization and manual verification of the FELINE detections (see Figure 2).
 
