+++ joss/paper.bib
@@ -3,7 +3,7 @@
 @ARTICLE{HerenzE_17a,
        author = {{Herenz}, Edmund Christian and {Wisotzki}, Lutz},
         title = "{LSDCat: Detection and cataloguing of emission-line sources in integral-field spectroscopy datacubes}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
      keywords = {methods: data analysis, techniques: imaging spectroscopy, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2017,
         month = jun,
@@ -41,7 +41,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Bacon+22,
        author = {{Bacon}, Roland and {Brinchmann}, Jarle and {Conseil}, Simon and {Maseda}, Michael and {Nanayakkara}, Themiya and {Wendt}, Martin and {Bacher}, Raphael and {Mary}, David and {Weilbacher}, Peter M. and {Krajnovi{\'c}}, Davor and {Boogaard}, Leindert and {Bouch{\'e}}, Nicolas and {Contini}, Thierry and {Epinat}, Beno{\^\i}t and {Feltre}, Anna and {Guo}, Yucheng and {Herenz}, Christian and {Kollatschny}, Wolfram and {Kusakabe}, Haruka and {Leclercq}, Floriane and {Michel-Dansac}, L{\'e}o and {Pello}, Roser and {Richard}, Johan and {Roth}, Martin and {Salvignol}, Gregory and {Schaye}, Joop and {Steinmetz}, Matthias and {Tresse}, Laurence and {Urrutia}, Tanya and {Verhamme}, Anne and {Vitte}, Eloise and {Wisotzki}, Lutz and {Zoutendijk}, Sebastiaan L.},
         title = "{The MUSE Hubble Ultra Deep Field surveys: Data release II}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
      keywords = {galaxies: high-redshift, galaxies: distances and redshifts, techniques: imaging spectroscopy, catalogs, Astrophysics - Astrophysics of Galaxies},
          year = 2023,
         month = feb,
@@ -91,7 +91,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Langan2023,
        author = {{Langan}, Ivanna and {Zabl}, Johannes and {Bouch{\'e}}, Nicolas F. and {Ginolfi}, Michele and {Popping}, Gerg{\"o} and {Schroetter}, Ilane and {Wendt}, Martin and {Schaye}, Joop and {Boogaard}, Leindert and {Freundlich}, Jonathan and {Richard}, Johan and {Matthee}, Jorryt and {Mercier}, Wilfried and {Contini}, Thierry and {Guo}, Yucheng and {Cherrey}, Maxime},
         title = "{MusE GAs FLOw and Wind (MEGAFLOW) IX. The impact of gas flows on the relations between the mass, star formation rate, and metallicity of galaxies}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {galaxies: abundances, galaxies: evolution, galaxies: formation, galaxies: ISM, (galaxies:) quasars: absorption lines, Astrophysics - Astrophysics of Galaxies},
          year = 2023,
         month = may,
@@ -109,7 +109,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Cherrey2024,
[…]
