+++ joss/paper.bib
@@ -145,12 +145,12 @@ archivePrefix = {arXiv},
 @ARTICLE{Bouche2024,
     author = {{Bouch{\'e}, Nicolas F. and {Wendt}, Martin and {Zabl}, Johannes and {Cherrey}, Maxime and {Schroetter}, Ilane and {Langan}, Ivanna and {Muzahid}, Sowgat and {Schaye}, Joop and {Epinat}, Benoît and {Wisotzki}, Lutz and {Contini}, Thierry and {Richard}, Johan and {Bacon}, Roland and {Weilbacher}, Peter M.},
     title = {MusE GAs FLOw and Wind (MEGAFLOW) - XII. Rationale and design of a Mg II survey of the cool circum-galactic medium with MUSE and UVES: The MEGAFLOW Survey},
-    DOI= "10.1051/0004-6361/202451093",
-    url= "https://doi.org/10.1051/0004-6361/202451093",
+    DOI= {"10.1051/0004-6361/202451093"},
+    url= {"https://doi.org/10.1051/0004-6361/202451093"},
     journal = {A&A},
     year = 2025,
-    volume = 694,
-    pages = "A67",
+    volume = {694},
+    pages = {"A67"},
 }
 
 @ARTICLE{Mary2020,
@@ -170,7 +170,7 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2020A&A...635A.194M},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-        
+
 @ARTICLE{Bertin1996,
        author = {{Bertin}, E. and {Arnouts}, S.},
         title = "{SExtractor: Software for source extraction.}",
