+++ joss/paper.bib
@@ -150,7 +150,7 @@ archivePrefix = {arXiv},
     journal = {A&A},
     year = 2025,
     volume = {694},
-    pages = {"A67"},
+    pages = {A67}
 }
 
 @ARTICLE{Mary2020,
