+++ joss/paper.bib
@@ -146,7 +146,6 @@ archivePrefix = {arXiv},
     author = {{Bouch{\'e}, Nicolas F. and {Wendt}, Martin and {Zabl}, Johannes and {Cherrey}, Maxime and {Schroetter}, Ilane and {Langan}, Ivanna and {Muzahid}, Sowgat and {Schaye}, Joop and {Epinat}, Benoît and {Wisotzki}, Lutz and {Contini}, Thierry and {Richard}, Johan and {Bacon}, Roland and {Weilbacher}, Peter M.},
     title = "{MusE GAs FLOw and Wind (MEGAFLOW) - XII. Rationale and design of a Mg II survey of the cool circum-galactic medium with MUSE and UVES: The MEGAFLOW Survey}",
     doi = {10.1051/0004-6361/202451093},
-    url = {https://doi.org/10.1051/0004-6361/202451093},
     journal = {A&A},
     year = 2025,
     volume = {694},
