+++ joss/paper.bib
@@ -149,7 +149,8 @@ archivePrefix = {arXiv},
       journal = {Astronomy and Astrophysics},
          year = 2025,
        volume = {694},
-        pages = {A67}
+        pages = {A67},
+          eid = {A67}
 }
 
 @ARTICLE{Mary2020,
