### 2025-03-06 96528c3
minor edits in the paper and bib
