+++ joss/paper.bib
@@ -56,16 +56,23 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-
-@ARTICLE{Bacon+14,
-       author = {{Bacon}, R. and {Vernet}, J. and {Borisova}, E. and {Bouch{\'e}}, N. and {Brinchmann}, J. and {Carollo}, M. and {Carton}, D. and {Caruana}, J. and {Cerda}, S. and {Contini}, T. and {Franx}, M. and {Girard}, M. and {Guerou}, A. and {Haddad}, N. and {Hau}, G. and {Herenz}, C. and {Herrera}, J.~C. and {Husemann}, B. and {Husser}, T. -O. and {Jarno}, A. and {Kamann}, S. and {Krajnovic}, D. and {Lilly}, S. and {Mainieri}, V. and {Martinsson}, T. and {Palsa}, R. and {Patricio}, V. and {P{\'e}contal}, A. and {Pello}, R. and {Piqueras}, L. and {Richard}, J. and {Sandin}, C. and {Schroetter}, I. and {Selman}, F. and {Shirazi}, M. and {Smette}, A. and {Soto}, K. and {Streicher}, O. and {Urrutia}, T. and {Weilbacher}, P. and {Wisotzki}, L. and {Zins}, G.},
-        title = "{MUSE Commissioning}",
-      journal = {The Messenger},
-         year = 2014,
-        month = sep,
-       volume = {157},
-        pages = {13-16},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2014Msngr.157...13B},
+@INPROCEEDINGS{2010SPIE.7735E..08B,
+       author = {{Bacon}, R. and {Accardo}, M. and {Adjali}, L. and {Anwand}, H. and {Bauer}, S. and {Biswas}, I. and {Blaizot}, J. and {Boudon}, D. and {Brau-Nogue}, S. and {Brinchmann}, J. and {Caillier}, P. and {Capoani}, L. and {Carollo}, C.~M. and {Contini}, T. and {Couderc}, P. and {Daguis{\'e}}, E. and {Deiries}, S. and {Delabre}, B. and {Dreizler}, S. and {Dubois}, J. and {Dupieux}, M. and {Dupuy}, C. and {Emsellem}, E. and {Fechner}, T. and {Fleischmann}, A. and {Fran{\c{c}}ois}, M. and {Gallou}, G. and {Gharsa}, T. and {Glindemann}, A. and {Gojak}, D. and {Guiderdoni}, B. and {Hansali}, G. and {Hahn}, T. and {Jarno}, A. and {Kelz}, A. and {Koehler}, C. and {Kosmalski}, J. and {Laurent}, F. and {Le Floch}, M. and {Lilly}, S.~J. and {Lizon}, J. -L. and {Loupias}, M. and {Manescau}, A. and {Monstein}, C. and {Nicklas}, H. and {Olaya}, J. -C. and {Pares}, L. and {Pasquini}, L. and {P{\'e}contal-Rousset}, A. and {Pell{\'o}}, R. and {Petit}, C. and {Popow}, E. and {Reiss}, R. and {Remillieux}, A. and {Renault}, E. and {Roth}, M. and {Rupprecht}, G. and {Serre}, D. and {Schaye}, J. and {Soucail}, G. and {Steinmetz}, M. and {Streicher}, O. and {Stuik}, R. and {Valentin},, H. and {Vernet}, J. and {Weilbacher}, P. and {Wisotzki}, L. and {Yerle}, N.},
+        title = "{The MUSE second-generation VLT instrument}",
[…]
