### 2025-03-07 #7528 @enthusi
@ivastar
As far as I can tell, all comments have been addressed, and the github release is linked into Zenodo:
https://doi.org/10.5281/zenodo.14987461
