### 2025-03-07 #7528 @ivastar
> * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

@enthusi can you please update the Zenodo title so it matches the paper?

Also, the version seems to be 1.1.0. Is this correct?
