### 2025-03-07 #7528 @enthusi
@ivastar Sorry for the title mismatch. I read that part wrong. It matches now.
https://doi.org/10.5281/zenodo.14987461
And we opted for 1.1.0 now that the software review is done. I hope that is ok. Nothing has changed since then (but compared to the initial submission 1.0.0).
