### 2025-03-11 #7528 @enthusi
@ivastar regarding the missing DOI for the cited MPDAF software package: there appears to be none for the version/rewrite.
