### 2025-03-11 5c85736
Minor updates to paper.md
