### 2025-03-11 #7528 @ivastar
@enthusi, I see that there is no DOI, that's ok.

I made a couple of cosmetic changes and made a pull request. Please review and merge if you agree and then I am ready to move this submission to the Editor in Chief for final review.
