### 2025-03-11 9905e55
Remove AAS journal info (#34)
