### 2025-03-15 ee9bc82
Initial editorial tweaks for JOSS paper and documentation typos
