+++ joss/paper.md
@@ -3,9 +3,8 @@ title: 'FELINE: A Tool to Detect Emission Line Galaxies in 3D Data'
 tags:
   - Python
   - astronomy
-  - ...
 authors:
-  - name: Martin Wendt^[mwendt@astro.physik.uni-potsdam.de]
+  - name: Martin Wendt
     orcid: 0000-0001-5020-9994
     affiliation: 1 # (Multiple affiliations must be quoted)
   - name: Marvin Henschel
@@ -34,19 +33,18 @@ Simple flux level peak detection algorithms based on thresholding are prone to e
 The *F*ind *E*mission *LINE*s tool ``FELINE`` combines a fully parallelized galaxy line template matching with the matched filter approach for individual emission features of LSDcat [@HerenzE_17a; @herenz2023].
 This transformation cross-correlates the dataset with a template that ideally matches the properties of the signal to be detected. Here, the images are cross-correlated with a 2D template for the expected light distribution of the sources to be detected and the same is applied in the spectral direction. The filter maximizes the signal-to-noise ratio (S/N) of a source that is optimally represented by the template while still producing a significant S/N enhancement for similar signals.
 
-The VLT/MUSE [@Bacon+10] 3D spectrograph creates $\sim$ 90,000 medium-resolution spectra arranged in a 300 $\times$ 300 spatial grid.
-These data cubes have typical sizes of 3-6 GiB per exposure, the sheer volume of data
+The VLT/MUSE [@Bacon+10] 3D spectrograph creates ~90,000 medium-resolution spectra arranged in a 300×300 spatial grid.
+These data cubes have typical sizes of 3-6 GiB per exposure.  The sheer volume of data
 requires automated processes to support the scientists.
-￼
 The ``FELINE`` algorithm evaluates the likelihood of emission lines at specific positions in each spectrum of the data cube. It does this by probing all possible combinations of up to 14 typical emission features, including Hα, Hβ, Hγ, Hδ, [OII], [OIII], [NII], [SII], and [NeIII], for the redshift range of interest (0.4 < z < 1.4). This extensive analysis leads to approximately $2.3\times 10^{11}$  iterations.
 
-There are generally two approaches to source identification. Image based, such as Source Extractor [@Bertin1996] and its many
-derivatives or emission line based, such as, the ORIGIN code [@Mary2020] or the previously mentioned LSDcat [@HerenzE_17a; @herenz2023].
-The former group analyzes white or narrow band images and is prone to miss objects with minimal or no continuum flux.
-On the other hand specialized software that focuses on detecting emission features is generally targeted at individual lines, like Lyman-$\alpha$. In contrast, the ``FELINE`` tool evaluates a physically connected set of emission lines simultaneously
+There are generally two approaches to source identification. Image-based, such as Source Extractor [@Bertin1996] and its many
+derivatives or emission-line-based, such as the ORIGIN code [@Mary2020] or the previously mentioned LSDcat [@HerenzE_17a; @herenz2023].
+The former group analyzes white or narrow-band images and is prone to miss objects with minimal or no continuum flux.
+On the other hand specialized software that focuses on detecting emission features is generally targeted at individual lines, like Lyman-$\alpha$. In contrast, the ``FELINE`` tool evaluates a physically-connected set of emission lines simultaneously
 and has been successfully tested as an automated identification system, providing rapid results for 3D data cubes that can be easily reviewed by humans.
-In @Bouche2024 we use a dual galaxy identiﬁcation process based both on continuum (using Source Extractor) and on
-emission lines using FELINE. Overall, we find that up to one third of our detected galaxies have no continuum
+In @Bouche2024 we use a dual galaxy identification process based both on continuum (using Source Extractor) and on
[…]
