### 2025-03-15 56e081c
consistently name package FELINE
