### 2025-03-15 #7528 @warrickball
Hi @enthusi! I've just opened [a pull request](https://github.com/enthusi/feline/pull/35) with some editorial tweaks. If these are okay, let me know when you've merged them and I'll give the publication pipeline another test run.
