### 2025-03-16 #7528 @enthusi
@warrickball Thank you for catching those typos and of course for streamlining the text.
I have merged your pull request happily.
