### 2025-03-16 #7528 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
